Update routes now follow a MVC format
